// schema.d.ts
export interface operations {
// ...
    "person-combined-credits": {
        parameters: {
            query?: {
                language?: string;
            };
            header?: never;
            path: {
                person_id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        cast?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /3h1JZGDhZ8nzxdgvkxha0qBqi05.jpg */
                            backdrop_path?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 13
                             */
                            id: number;
                            /** @example en */
                            original_language?: string;
                            /** @example Forrest Gump */
                            original_title?: string;
                            /** @example A man with a low IQ has accomplished great things in his life and been present during significant historic events—in each case, far exceeding what anyone imagined he could do. But despite all he has achieved, his one true love eludes him. */
                            overview?: string;
                            /**
                             * @default 0
                             * @example 62.225
                             */
                            popularity: number;
                            /** @example /arw2vcBveWOVZr6pxd9XTd1TdQa.jpg */
                            poster_path?: string;
                            /** @example 1994-06-23 */
                            release_date?: string;
                            /** @example Forrest Gump */
                            title?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 8.481
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 24535
                             */
                            vote_count: number;
                            /** @example Forrest Gump */
                            character?: string;
                            /** @example 52fe420ec3a36847f800074f */
                            credit_id?: string;
                            /**
                             * @default 0
                             * @example 0
                             */
                            order: number;
                            /** @example movie */
                            media_type?: string;
                        }[];
                        crew?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /tx3uj8GPWf5pzb0gWATJ4bokNHI.jpg */
                            backdrop_path?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 87061
                             */
                            id: number;
                            /** @example fr */
                            original_language?: string;
                            /** @example Le Voyage extraordinaire */
                            original_title?: string;
                            /** @example An account of the extraordinary life of film pioneer Georges Méliès (1861-1938) and the amazing story of the copy in color of his masterpiece “A Trip to the Moon” (1902), unexpectedly found in Spain and restored thanks to the heroic efforts of a group of true cinema lovers. */
                            overview?: string;
                            /**
                             * @default 0
                             * @example 6.007
                             */
                            popularity: number;
                            /** @example /zHNNT9gfiGsuadR6x38KYOp6ekq.jpg */
                            poster_path?: string;
                            /** @example 2011-12-08 */
                            release_date?: string;
                            /** @example The Extraordinary Voyage */
                            title?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 7.6
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 47
                             */
                            vote_count: number;
                            /** @example 5d818a63d34eb3002c4f8fea */
                            credit_id?: string;
                            /** @example Crew */
                            department?: string;
                            /** @example Thanks */
                            job?: string;
                            /** @example movie */
                            media_type?: string;
                        }[];
                        /**
                         * @default 0
                         * @example 31
                         */
                        id: number;
                    };
                };
            };
        };
    };
// ...
}